pyproject.toml
[project]
name = "mycv"
version = "1.0.0"
description = "AI-powered CV/Resume builder"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.100.0",
    "pydantic>=2.0",
    "uvicorn>=0.32.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
default-groups = ["dev"]
